setup_env.sh argument count

+++ scripts/setup_env.sh
@@ -3,8 +3,10 @@ set -euxo pipefail
 
 NAMESPACE="ingress-nginx"
 
-if [ "$1" == "delete" ]; then
-    k3d cluster delete || true
+if [ "$#" -eq 1 ]; then
+    if [ "$1" == "delete" ]; then
+        k3d cluster delete || true
+    fi
 fi
 
 k3d cluster create --wait --no-lb -p "32080:32080@server[*]" -p "32443:32443@server[*]"
